Cleanup, code extract and reuse
