+++ src/pages/Facility/services/inventory/externalSupply/deliveryOrder/DeliveryOrderShow.tsx
@@ -399,7 +399,7 @@ export function DeliveryOrderShow({
       title={t("delivery_order_details")}
       hideTitleOnPage
       shortCutContext="facility:inventory:delivery"
-      className="max-w-7xl mx-auto"
+      className="max-w-[1600px] mx-auto"
     >
       <div className="space-y-6">
         <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2">
