+++ src/pages/Facility/services/inventory/externalSupply/utils/inventoryUtils.ts
@@ -1,4 +1,8 @@
-import { MonetaryComponentType } from "@/types/base/monetaryComponent/monetaryComponent";
+import {
+  calculateItemTotal,
+  MonetaryComponentType,
+} from "@/types/base/monetaryComponent/monetaryComponent";
+import { getComponentsFromChargeItem } from "@/types/billing/chargeItem/chargeItem";
 import { SupplyDeliveryRead } from "@/types/inventory/supplyDelivery/supplyDelivery";
 
 export const getInventoryBasePath = (
@@ -27,48 +31,32 @@ export const getInventoryBasePath = (
   }
 };
 
+/**
+ * Calculate total price for a supply delivery
+ * Uses shared utilities for component extraction and total calculation
+ */
 export function calculateTotal(
   delivery: SupplyDeliveryRead,
   internal: boolean,
 ): number {
-  const priceComponents = internal
+  const chargeItemDef = internal
     ? delivery.supplied_inventory_item?.product?.charge_item_definition
-        ?.price_components
-    : delivery.supplied_item?.charge_item_definition?.price_components;
+    : delivery.supplied_item?.charge_item_definition;
 
-  if (!priceComponents) return 0;
+  if (!chargeItemDef) return 0;
 
-  const baseComponent = priceComponents.find(
-    (c) => c.monetary_component_type === MonetaryComponentType.base,
+  const baseComponents = getComponentsFromChargeItem(
+    chargeItemDef,
+    MonetaryComponentType.base,
   );
-  const basePrice = baseComponent?.amount
-    ? parseFloat(baseComponent.amount)
+  const basePrice = baseComponents[0]?.amount
+    ? parseFloat(baseComponents[0].amount)
     : 0;
-  const quantity = delivery.supplied_item_quantity || 1;
 
-  let total = basePrice * quantity;
-
-  // Apply taxes
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.tax)
-    .forEach((tax) => {
-      if (tax.factor) {
-        total += basePrice * quantity * (tax.factor / 100);
-      } else if (tax.amount) {
-        total += parseFloat(tax.amount);
-      }
-    });
-
-  // Apply discounts
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.discount)
-    .forEach((discount) => {
-      if (discount.factor) {
-        total -= basePrice * quantity * (discount.factor / 100);
-      } else if (discount.amount) {
-        total -= parseFloat(discount.amount);
-      }
-    });
-
-  return total;
+  return calculateItemTotal(
+    basePrice,
+    delivery.supplied_item_quantity || 1,
+    getComponentsFromChargeItem(chargeItemDef, MonetaryComponentType.tax),
+    getComponentsFromChargeItem(chargeItemDef, MonetaryComponentType.discount),
+  );
 }
