+++ src/pages/Facility/settings/chargeItemDefinitions/ChargeItemDefinitionForm.tsx
@@ -2,9 +2,10 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { CheckIcon, Loader2 } from "lucide-react";
 import { navigate } from "raviger";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
+import { toast } from "sonner";
 import * as z from "zod";
 
 import { Button } from "@/components/ui/button";
@@ -32,20 +33,20 @@ import {
 } from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
 
-import { CompactConditionEditor } from "@/components/Billing/CompactConditionEditor";
+import { MonetaryComponentSelector } from "@/components/Billing/MonetaryComponentSelector";
 import Loading from "@/components/Common/Loading";
 import { ResourceCategoryPicker } from "@/components/Common/ResourceCategoryPicker";
 
 import { cn } from "@/lib/utils";
+
 import { CodeSchema } from "@/types/base/code/code";
 import {
   ConditionForm,
   conditionSchema,
-  Metrics,
 } from "@/types/base/condition/condition";
 import {
   MonetaryComponent,
-  MonetaryComponentRead,
+  monetaryComponentIsEqual,
   MonetaryComponentType,
 } from "@/types/base/monetaryComponent/monetaryComponent";
 import { ResourceCategoryResourceType } from "@/types/base/resourceCategory/resourceCategory";
@@ -62,37 +63,7 @@ import chargeItemDefinitionApi from "@/types/billing/chargeItemDefinition/charge
 import facilityApi from "@/types/facility/facilityApi";
 import mutate from "@/Utils/request/mutate";
 import query from "@/Utils/request/query";
-import { Check, ChevronDown, Component, Search, Trash2 } from "lucide-react";
-
-import { Badge } from "@/components/ui/badge";
-import { Checkbox } from "@/components/ui/checkbox";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-
-import RadioInput from "@/components/ui/RadioInput";
 import { generateSlug } from "@/Utils/utils";
-import { toast } from "sonner";
-
-interface MonetaryComponentSelectorProps {
-  title: string;
-  components: MonetaryComponentRead[];
-  selectedComponents: MonetaryComponent[];
-  onComponentToggle: (
-    component: MonetaryComponent,
-    selected: boolean,
-    type: MonetaryComponentType,
-  ) => void;
-  onConditionsChange: (
-    component: MonetaryComponent,
-    conditions: ConditionForm[],
-  ) => void;
-  type: MonetaryComponentType;
-  availableMetrics: Metrics[];
-  className?: string;
-}
 
 // Schema factory for a single price component
 const createPriceComponentSchema = (
@@ -119,427 +90,6 @@ interface ChargeItemDefinitionFormProps {
   isUpdate?: boolean;
   onSuccess?: (chargeItemDefinition: ChargeItemDefinitionRead) => void;
   onCancel?: () => void;
-}
-
-const monetaryComponentIsEqual = <T extends MonetaryComponent>(a: T, b: T) => {
-  return (
-    a.monetary_component_type === b.monetary_component_type &&
-    a.code?.code === b.code?.code &&
-    a.code?.system === b.code?.system &&
-    a.code?.display === b.code?.display
-  );
-};
-
-// Component for monetary component selection with autocomplete
-export function MonetaryComponentSelector({
-  title,
-  components,
-  selectedComponents,
-  onComponentToggle,
-  onConditionsChange,
-  type,
-  availableMetrics,
-  className = "",
-}: MonetaryComponentSelectorProps) {
-  const { t } = useTranslation();
-  const [isOpen, setIsOpen] = useState(false);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [tempSelectedComponents, setTempSelectedComponents] = useState<
-    MonetaryComponent[]
-  >([]);
-
-  // Initialize temp selections when dropdown opens
-  useEffect(() => {
-    if (isOpen) {
-      setTempSelectedComponents(selectedComponents);
-    }
-  }, [isOpen, selectedComponents]);
-
-  const getComponentValue = (component: MonetaryComponent) => {
-    return component.factor ?? component.amount ?? 0;
-  };
-
-  const isSameAmountOrFactor = (
-    component: MonetaryComponent,
-    otherComponent: MonetaryComponent,
-  ) => {
-    return (
-      (component.factor != null &&
-        component.factor === otherComponent.factor) ||
-      (component.amount != null && component.amount === otherComponent.amount)
-    );
-  };
-
-  const isComponentSelected = (
-    component: MonetaryComponentRead,
-    selectedComponents: MonetaryComponent[],
-  ) =>
-    selectedComponents.some(
-      (c) =>
-        monetaryComponentIsEqual(c, component) &&
-        isSameAmountOrFactor(c, component),
-    );
-
-  // Group components by code
-  const groupedComponents = components.reduce<
-    Record<string, MonetaryComponentRead[]>
-  >((acc, component) => {
-    const key = component.code?.code;
-    if (key) {
-      (acc[key] ||= []).push(component);
-    }
-    return acc;
-  }, {});
-
-  const groupComponents: Record<string, MonetaryComponentRead[]> = {};
-  const nonGroupComponents: MonetaryComponentRead[] = [];
-
-  Object.entries(groupedComponents).forEach(([key, comps]) => {
-    if (comps.length > 1) {
-      groupComponents[key] = comps;
-    } else {
-      nonGroupComponents.push(comps[0]);
-    }
-  });
-
-  const filteredGroupComponents = Object.entries(groupComponents).reduce<
-    Record<string, MonetaryComponentRead[]>
-  >((acc, [key, comps]) => {
-    const filteredComps = comps.filter(
-      (component) =>
-        component.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
-        component.code?.code?.toLowerCase().includes(searchQuery.toLowerCase()),
-    );
-
-    if (filteredComps.length > 0) {
-      acc[key] = filteredComps;
-    }
-
-    return acc;
-  }, {});
-
-  const filteredNonGroupComponents = nonGroupComponents.filter(
-    (component) =>
-      component.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
-      component.code?.code?.toLowerCase().includes(searchQuery.toLowerCase()),
-  );
-
-  const handleTempToggle = (
-    component: MonetaryComponentRead,
-    checked: boolean,
-  ) => {
-    if (checked) {
-      const newComponent: MonetaryComponent = {
-        ...component,
-        monetary_component_type: type,
-        factor: component.factor != null ? component.factor : undefined,
-        amount:
-          component.factor != null ? undefined : String(component.amount || 0),
-        conditions: [],
-      };
-      setTempSelectedComponents((prev) => [...prev, newComponent]);
-    } else {
-      setTempSelectedComponents((prev) =>
-        prev.filter((c) => !monetaryComponentIsEqual(c, component)),
-      );
-    }
-  };
-
-  const handleDone = () => {
-    // Remove all current selections first
-    selectedComponents.forEach((component) => {
-      onComponentToggle(component, false, type);
-    });
-
-    // Add all temp selections
-    tempSelectedComponents.forEach((component) => {
-      onComponentToggle(component, true, type);
-    });
-
-    setIsOpen(false);
-    setSearchQuery("");
-  };
-
-  const handleCancel = () => {
-    setIsOpen(false);
-    setSearchQuery("");
-    setTempSelectedComponents([]);
-  };
-
-  const handleRadioButtonChange = (groupKey: string, selected: string) => {
-    // If selected is empty, it means the user deselected the radio button
-    if (!selected) {
-      // Remove any component from this group
-      setTempSelectedComponents((prev) =>
-        prev.filter((c) => c.code?.code !== groupKey),
-      );
-      return;
-    }
-
-    const group = groupComponents[groupKey];
-    if (!group) return;
-
-    // Find the component with the matching value
-    const selectedComponent = group.find(
-      (component) => getComponentValue(component).toString() === selected,
-    );
-
-    if (!selectedComponent) return;
-
-    // Remove any existing selection from this group and add the new one
-    setTempSelectedComponents((prev) => {
-      // Remove any component from the same group (by code)
-      const filtered = prev.filter((c) => c.code?.code !== groupKey);
-      return [...filtered, selectedComponent];
-    });
-  };
-
-  const renderGroupCheckList = (
-    groupComponents: Record<string, MonetaryComponentRead[]>,
-  ) => {
-    if (Object.keys(groupComponents).length === 0) {
-      return <></>;
-    }
-    return (
-      <div className="flex flex-col gap-2">
-        {Object.entries(groupComponents).map(([key, components]) => {
-          const selectedInGroup = components.find((component) =>
-            isComponentSelected(component, tempSelectedComponents),
-          );
-          const selectedValue = selectedInGroup
-            ? `${getComponentValue(selectedInGroup)}`
-            : "";
-
-          const radioOptions = components.map((component) => ({
-            label: `${getComponentValue(component)} ${component.factor != null ? "%" : "₹"}`,
-            value: `${getComponentValue(component)}`,
-          }));
-
-          return (
-            <div key={`${key}`} className="flex flex-col gap-2">
-              <div className="flex items-center gap-2 p-2">
-                <Component
-                  className="size-4 text-black/80"
-                  strokeWidth={1.25}
-                />
-                <div className="text-sm font-semibold text-gray-900 uppercase">
-                  {key}
-                </div>
-              </div>
-              <RadioInput
-                value={selectedValue}
-                onValueChange={(value: string) =>
-                  handleRadioButtonChange(key, value)
-                }
-                options={radioOptions}
-                className="flex flex-row gap-1 justify-end mr-2"
-              />
-            </div>
-          );
-        })}
-      </div>
-    );
-  };
-
-  const renderCheckList = (
-    listComponents: MonetaryComponentRead[],
-    isGroup = false,
-  ) => {
-    if (listComponents.length === 0) {
-      return <></>;
-    }
-    return listComponents.map((component, idx) => {
-      const isSelected = isComponentSelected(component, tempSelectedComponents);
-      return (
-        <div
-          key={`${component.title}-${component.code?.code || idx}`}
-          className={cn(
-            "flex items-center space-x-3 p-2 hover:bg-gray-50 rounded",
-          )}
-        >
-          <Checkbox
-            checked={isSelected}
-            onCheckedChange={(checked) =>
-              handleTempToggle(component, checked as boolean)
-            }
-            className={cn(isGroup && "aria-hidden")}
-          />
-          <div className="flex flex-row justify-between items-center flex-1 min-w-0">
-            <div className="text-sm font-medium text-gray-900">
-              {component.code?.display}
-            </div>
-            <div className="flex flex-row items-center gap-2 h-10">
-              {getComponentValue(component)}
-              <span className="text-gray-500">
-                {component.factor != null ? "%" : "₹"}
-              </span>
-            </div>
-          </div>
-        </div>
-      );
-    });
-  };
-
-  return (
-    <div className={cn("space-y-1", className)}>
-      {/* Selected Components Section - Only for Discounts */}
-      {type === MonetaryComponentType.discount &&
-        selectedComponents.length > 0 && (
-          <div className="space-y-1 mb-2">
-            <p className="text-sm font-medium text-gray-700">
-              {t("selected")} {title.toLowerCase()}
-            </p>
-
-            {selectedComponents.map((component, idx) => {
-              const componentRead = components.find((c) =>
-                monetaryComponentIsEqual(c, component),
-              );
-
-              return (
-                <div
-                  key={`selected-${componentRead?.title}-${componentRead?.code?.code || idx}`}
-                  className="p-3 rounded-lg bg-white border border-gray-200 transition-colors"
-                >
-                  <div className="flex items-center justify-between border-b pb-2">
-                    <div>
-                      <div className="font-medium text-md">
-                        {idx + 1}. {componentRead?.code?.display} -{" "}
-                        {getComponentValue(component)}
-                        {component.factor != null ? "%" : "₹"}
-                      </div>
-                    </div>
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="sm"
-                      onClick={() => onComponentToggle(component, false, type)}
-                      className="h-6 w-6 p-0"
-                    >
-                      <Trash2 className="size-4" />
-                    </Button>
-                  </div>
-
-                  {/* Condition editor for discount components */}
-                  <CompactConditionEditor
-                    conditions={
-                      component.conditions?.map((condition) => ({
-                        ...condition,
-                        _conditionType: `${condition.metric}_${condition.operation}`,
-                      })) || []
-                    }
-                    availableMetrics={availableMetrics}
-                    onChange={(conditions) =>
-                      onConditionsChange(
-                        { ...component, monetary_component_type: type },
-                        conditions,
-                      )
-                    }
-                    className="mt-3"
-                  />
-                </div>
-              );
-            })}
-          </div>
-        )}
-
-      <div className="flex items-center justify-between">
-        <div>
-          <p className="text-sm text-gray-900">{title}</p>
-        </div>
-      </div>
-
-      {/* Trigger Area */}
-      <Popover open={isOpen} onOpenChange={setIsOpen}>
-        <PopoverTrigger asChild>
-          <div className="bg-white border rounded-md p-3 cursor-pointer hover:border-gray-400 transition-colors min-h-[44px] flex items-center justify-between">
-            <div className="flex items-center gap-2 flex-wrap">
-              {selectedComponents.length === 0 ? (
-                <span className="text-gray-500 text-sm">
-                  {t(
-                    type === MonetaryComponentType.tax
-                      ? "add_tax"
-                      : "add_discount",
-                  )}
-                </span>
-              ) : type === MonetaryComponentType.tax ? (
-                // For taxes, show badges in trigger
-                <>
-                  {selectedComponents.slice(0, 3).map((component, idx) => {
-                    const value = getComponentValue(component);
-                    const suffix = component.factor != null ? "%" : "₹";
-                    const display = component.code?.display;
-                    return (
-                      <Badge
-                        key={`${component.code?.code}-${idx}`}
-                        variant="secondary"
-                        className="text-xs p-1 rounded-sm"
-                      >
-                        {display} @ {value}
-                        {suffix}
-                      </Badge>
-                    );
-                  })}
-                  {selectedComponents.length > 3 && (
-                    <Badge variant="secondary" className="text-xs">
-                      +{selectedComponents.length - 3} {t("more")}
-                    </Badge>
-                  )}
-                </>
-              ) : (
-                // For discounts, just show count
-                <span className="text-gray-700 text-sm">
-                  {selectedComponents.length} {t("selected")}
-                </span>
-              )}
-            </div>
-            <ChevronDown className="size-4 text-gray-400" />
-          </div>
-        </PopoverTrigger>
-
-        <PopoverContent className="w-68 p-0" align="start">
-          <div className="p-3 border-b">
-            <div className="relative">
-              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 size-4 text-gray-400" />
-              <Input
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                placeholder={t(
-                  type === MonetaryComponentType.tax
-                    ? "search_for_tax_code"
-                    : "search_for_discount_code",
-                )}
-                className="pl-10"
-              />
-            </div>
-          </div>
-
-          <div className="max-h-[30vh] overflow-y-auto p-2">
-            {renderGroupCheckList(filteredGroupComponents)}
-            {renderCheckList(filteredNonGroupComponents)}
-          </div>
-
-          <div className="p-3 border-t flex gap-2">
-            <Button
-              onClick={handleCancel}
-              variant="outline"
-              size="sm"
-              className="flex-1"
-            >
-              {t("cancel")}
-            </Button>
-            <Button
-              onClick={handleDone}
-              size="sm"
-              className="flex-1 bg-green-600 hover:bg-green-700"
-            >
-              <Check className="size-4 mr-1" />
-              {t("done")}
-            </Button>
-          </div>
-        </PopoverContent>
-      </Popover>
-    </div>
-  );
 }
 
 export function ChargeItemDefinitionForm({
@@ -850,36 +400,32 @@ export function ChargeItemDefinitionForm({
   const getSelectedComponents = (type: MonetaryComponentType) =>
     priceComponents.filter((c) => c.monetary_component_type === type);
 
-  // Handle component selection
-  const handleComponentToggle = (
-    component: MonetaryComponent,
-    selected: boolean,
-    type: MonetaryComponentType = MonetaryComponentType.tax,
+  // Handle selection change from MonetaryComponentSelector
+  const handleSelectionChange = (
+    selectedComponents: MonetaryComponent[],
+    type: MonetaryComponentType,
   ) => {
     const currentComponents = form.getValues("price_components");
-    let newComponents: z.infer<typeof priceComponentSchema>[];
 
-    if (selected) {
-      newComponents = [
-        ...currentComponents,
-        {
-          ...component,
-          monetary_component_type: type,
-          factor: component.factor != null ? component.factor : undefined,
-          amount:
-            component.factor != null ? undefined : String(component.amount),
-          conditions:
-            component.conditions?.map((condition) => ({
-              ...condition,
-              _conditionType: `${condition.metric}_${condition.operation}`,
-            })) || [],
-        },
-      ];
-    } else {
-      newComponents = currentComponents.filter(
-        (c) => !monetaryComponentIsEqual(c, component),
-      );
-    }
+    // Remove all components of this type
+    const otherComponents = currentComponents.filter(
+      (c) => c.monetary_component_type !== type,
+    );
+
+    // Add all newly selected components
+    const newSelectedComponents = selectedComponents.map((component) => ({
+      ...component,
+      monetary_component_type: type,
+      factor: component.factor != null ? component.factor : undefined,
+      amount: component.factor != null ? undefined : String(component.amount),
+      conditions:
+        component.conditions?.map((condition) => ({
+          ...condition,
+          _conditionType: `${condition.metric}_${condition.operation}`,
+        })) || [],
+    }));
+
+    const newComponents = [...otherComponents, ...newSelectedComponents];
 
     form.setValue("price_components", newComponents, {
       shouldValidate: true,
@@ -1191,10 +737,10 @@ export function ChargeItemDefinitionForm({
               selectedComponents={getSelectedComponents(
                 MonetaryComponentType.tax,
               )}
-              onComponentToggle={handleComponentToggle}
-              onConditionsChange={handleComponentConditionsChange}
+              onSelectionChange={(components) =>
+                handleSelectionChange(components, MonetaryComponentType.tax)
+              }
               type={MonetaryComponentType.tax}
-              availableMetrics={availableMetrics}
               className={minimal ? "w-full" : ""}
             />
 
@@ -1204,9 +750,15 @@ export function ChargeItemDefinitionForm({
               selectedComponents={getSelectedComponents(
                 MonetaryComponentType.discount,
               )}
-              onComponentToggle={handleComponentToggle}
+              onSelectionChange={(components) =>
+                handleSelectionChange(
+                  components,
+                  MonetaryComponentType.discount,
+                )
+              }
               onConditionsChange={handleComponentConditionsChange}
               type={MonetaryComponentType.discount}
+              showConditionsEditor
               availableMetrics={availableMetrics}
               className={minimal ? "w-full" : ""}
             />
