+++ src/types/base/monetaryComponent/monetaryComponent.ts
@@ -28,3 +28,91 @@ export const MonetaryComponentOrder = {
   discount: 4,
   tax: 5,
 } as const satisfies Record<MonetaryComponentType, number>;
+
+// Utility functions for monetary component operations
+
+/**
+ * Compare two monetary components for equality based on type and code
+ */
+export function monetaryComponentIsEqual<T extends MonetaryComponent>(
+  a: T,
+  b: T,
+): boolean {
+  return (
+    a.monetary_component_type === b.monetary_component_type &&
+    a.code?.code === b.code?.code &&
+    a.code?.system === b.code?.system
+  );
+}
+
+/**
+ * Get the effective value of a monetary component (factor or amount)
+ */
+export function getComponentValue(
+  component: MonetaryComponent,
+): number | string {
+  return component.factor ?? component.amount ?? 0;
+}
+
+/**
+ * Check if two components have the same amount or factor value
+ */
+export function isSameAmountOrFactor(
+  a: MonetaryComponent,
+  b: MonetaryComponent,
+): boolean {
+  return (
+    (a.factor != null && a.factor === b.factor) ||
+    (a.amount != null && a.amount === b.amount)
+  );
+}
+
+/**
+ * Check if a component is selected in a list of components
+ */
+export function isComponentSelected(
+  component: MonetaryComponent,
+  selectedComponents: MonetaryComponent[],
+): boolean {
+  return selectedComponents.some(
+    (c) =>
+      monetaryComponentIsEqual(c, component) &&
+      isSameAmountOrFactor(c, component),
+  );
+}
+
+/**
+ * Calculate total price from base price, quantity, taxes, and discounts
+ */
+export function calculateItemTotal(
+  basePrice: number,
+  quantity: number,
+  taxComponents?: MonetaryComponent[],
+  discountComponents?: MonetaryComponent[],
+): number {
+  let total = basePrice * quantity;
+
+  // Apply taxes
+  if (taxComponents?.length) {
+    taxComponents.forEach((tax) => {
+      if (tax.factor) {
+        total += basePrice * quantity * (tax.factor / 100);
+      } else if (tax.amount) {
+        total += parseFloat(tax.amount);
+      }
+    });
+  }
+
+  // Apply discounts
+  if (discountComponents?.length) {
+    discountComponents.forEach((discount) => {
+      if (discount.factor) {
+        total -= basePrice * quantity * (discount.factor / 100);
+      } else if (discount.amount) {
+        total -= parseFloat(discount.amount);
+      }
+    });
+  }
+
+  return total;
+}
